Create blank files for use with set tracking
I am adding tracking sets like best of 5 to the chat bot.

These text files will have data written to them to be show on stream.

Additionally, the complete set can be logged to the set_log.txt

The rest of the code for this functionality will be pushed soon.