Create Streak logic, seperate file writing
A new set of logic has been created for a smash player on a streak in
the arena. This will write their streak number and their name to files
that can be used on stream.

Additionally, i have broken out the commands that write data to files so
that they can be used in other parts of the code. This increases code
reuse dramatically.