Modify commands for ease of use
The streak logic needed some work so that I only have to input 1 command
per streak end as opposed to 3.
The file writing for those proceses is easier as well.

The Arena commands were updated to have a non joinable message if it is
empty and set it to something that would trigger that message.

The set logic now writes to the files on reset much better than before.
It leaves the files in the same state that it found them.