Add Count feature, save dev work
A new counting feature has been added to the bot which I primarily will
use for MM2 but it could be used for anything.

Also I've saved some dev work relating to updating the updating of
streak arenas automatically. A lot more work must be done with this to
make it operational