Add V1 King of the Hill Functionality
V1 King of the Hill Functionality has been implemented. This is still a
work in progress as better handling of ties must be implemented.

Additionally, the streak must now be logged with an extra command which
is inconvienent. Ideally this could be done with one command to reduce
the streamers load while streaming / playing.

Lastly, the elitehelp command has not been updated with the new commands
related to this functionality. It will be added when this logic works
better.