Implement King of the Hill V2, fix bugs
I have implemented V2 of king of the hill that will output a leader
board to discord. This is the best way to resolve multiway ties.

A clear streaming command was created to delete the messages in the
streaming channel as well.
Elite help was updated with the new koth commands as well.