Continue work on Automated streaks
More work has been done to implement automated streaks.
Opencv & Jimp are being utilized to see who won a match when it returns
to the lobby.
Right now this only detects on sample images.
The next step is to parse out frames from a video and see who won.
And finally to parse it from the stream live. Code to save the stream to
an flv file is already comepleted.